# docs/static/specs/telegram-gateway-api.openapi.yaml
openapi: 3.1.0
info:
  title: TradingSystem Telegram Gateway API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    REST API for accessing Telegram messages, channels, and gateway status with TimescaleDB persistence.

    The Telegram Gateway API serves as the query interface for messages captured by the MTProto Gateway (Port 4007).
    It provides comprehensive CRUD operations for messages and channels, authentication management,
    and monitoring capabilities.

    **Key Features:**
    - Message retrieval with advanced filtering (channel, date range, status, search)
    - Channel CRUD operations
    - Telegram authentication management (interactive phone/code flow)
    - Gateway metrics and health monitoring
    - Failure queue management
    - Message reprocessing capabilities
    - Prometheus metrics export

    **Architecture:**
    - MTProto Gateway (Port 4007) → Captures messages from Telegram
    - TimescaleDB → Stores messages and channels
    - This API (Port 4010) → Query interface for downstream services

    **Rate Limiting:**
    - Global rate limit: 300 requests per minute per IP
    - Rate limit headers included in all responses

    **Authentication:**
    - Optional token-based authentication (disabled by default in development)
    - Set `TELEGRAM_GATEWAY_REQUIRE_TOKEN=true` to enable
    - Provide token via `X-Gateway-Token`, `X-API-Token`, or `Authorization: Bearer` headers
servers:
...
security: []
...
paths:
...
  /api/telegram-gateway/auth/start:
    post:
      tags: [Authentication]
      summary: Start authentication
      description: |
        Initiates Telegram authentication flow.

        **Flow:**
        1. Call this endpoint to start
        2. Poll `/auth/status` to check state
        3. Submit phone number via `/auth/input` when state is `waitingForPhone`
        4. Submit verification code via `/auth/input` when state is `waitingForCode`
        5. Submit 2FA password via `/auth/input` when state is `waitingForPassword` (if required)
      responses:
        '200':
          description: Authentication started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AuthStatus'
        '409':
          description: Authentication already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/telegram-gateway/auth/input:
    post:
      tags: [Authentication]
      summary: Submit authentication input
      description: |
        Submit phone number, verification code, or 2FA password during authentication flow.

        **Value Format:**
        - Phone: International format with + (e.g., +5511987654321)
        - Code: 5-digit code from Telegram
        - Password: Your 2FA password (if enabled)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - value
              properties:
                value:
                  type: string
                  description: Phone number, code, or password
              example:
                value: "+5511987654321"
      responses:
        '200':
          description: Input submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AuthStatus'
        '400':
          description: Missing value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: No authentication process running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/telegram-gateway/auth/cancel:
    post:
      tags: [Authentication]
      summary: Cancel authentication
      description: Cancels ongoing authentication process.
      responses:
        '200':
          description: Authentication cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AuthStatus'
        '409':
          description: No authentication process running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/telegram-gateway/queue:
    get:
      tags: [Gateway]
      summary: Get failure queue
      description: |
        Returns messages in the failure queue (messages that failed processing).

        **Use Case:** Monitor and manually review failed messages.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Max messages to return
      responses:
        '200':
          description: Failure queue
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
...
  /api/telegram-gateway/sync-messages:
    post:
      tags: [Sync]
      summary: Sync messages from Telegram
      description: |
        Triggers message synchronization from Telegram channels.

        **Note:** This proxies the request to the MTProto Gateway (Port 4007).
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      totalMessagesSynced:
                        type: integer
        '503':
          description: MTProto Gateway unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    AuthStatus:
      type: object
      properties:
        state:
          type: string
          enum: [idle, waitingForPhone, waitingForCode, waitingForPassword, authenticated, failed]
        message:
          type: string
        error:
          anyOf:
            - type: string
            - type: "null"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string